+++ test/test_map_class.py
@@ -1,3 +1,4 @@
+import glob
 import os
 import shutil
 from pathlib import Path
@@ -7,50 +8,23 @@ from hydra import compose, initialize
 
 import ctview.utils_pdal as utils_pdal
 from ctview.map_class import (
+    add_color_to_raster,
+    create_class_raster_raw,
     create_map_class_raster_with_postprocessing_color_and_hillshade,
-    step1_create_raster_brut,
-    step2_create_raster_fillgap,
-    step3_color_raster,
+    fill_gaps_raster,
 )
 
-# Test files
 OUTPUT_DIR = os.path.join("tmp", "map_class")
-INPUT_DIR = os.path.join("data", "las")
-INPUT_FILENAME = "test_data_0000_0000_LA93_IGN69_ground.las"
+INPUT_DIR = os.path.join("data", "las", "classee")
+INPUT_FILENAME = "test_data_77050_627755_LA93_IGN69.las"
+TILENAME = os.path.splitext(INPUT_FILENAME)[0]
 INPUT_FILE = os.path.join(INPUT_DIR, INPUT_FILENAME)
-IN_POINTS = utils_pdal.read_las_file(INPUT_FILE)  # tested
-FILENAME = "defaultname"
-VERBOSE = "suffix"
-LUT_FILE = os.path.join("LUT", "LUT_CLASS.txt")
+IN_POINTS = utils_pdal.read_las_file(INPUT_FILE)
 RASTER_DRIVER = "GTiff"
 
-# Config
 TILE_COORD_SCALE = 10
 TILE_WIDTH = 50
 BUFFER_SIZE = 10
-with initialize(version_base="1.2", config_path="../configs"):
-    # config is relative to a module
-    CONFIG = compose(
-        config_name="config_ctview",
-        overrides=[
-            f"io.tile_geometry.tile_coord_scale={TILE_COORD_SCALE}",
-            f"io.tile_geometry.tile_width={TILE_WIDTH}",
-            f"buffer.size={BUFFER_SIZE}",
-        ],
-    )
-
-
-# Attempted
-PATH1_EXPECTED = os.path.join(OUTPUT_DIR, "defaultname_raster.tif")
-PATH2_EXPECTED = os.path.join(OUTPUT_DIR, "defaultname_raster_fillgap.tif")
-PATH3_EXPECTED = os.path.join(OUTPUT_DIR, "defaultname_suffix.tif")
-
-EXPECTED_FILLGAP_COLOR = os.path.join(
-    OUTPUT_DIR, "CC_4_fgcolor", "test_data_0000_0000_LA93_IGN69_ground_raster_fillgap_color.tif"
-)
-EXPECTED_FILLGAP_COLOR_JPG = os.path.join(
-    OUTPUT_DIR, "CC_4_fgcolor", "test_data_0000_0000_LA93_IGN69_ground_raster_fillgap_color.jpg"
-)
 
 
 def setup_module(module):  # run before the first test
@@ -62,69 +36,173 @@ def setup_module(module):  # run before the first test
     os.makedirs(OUTPUT_DIR)
 
 
-def test_step1_create_raster_brut():
-    raster = step1_create_raster_brut(
-        in_points=IN_POINTS,
-        output_dir=OUTPUT_DIR,
-        output_filename=FILENAME,
-        res=1,
-        i=1,
-        output_extension=CONFIG.io.extension,
-        raster_driver=RASTER_DRIVER,
+def test_create_class_raster_raw():
+    output_file = Path(OUTPUT_DIR) / "create_class_raster_raw" / f"{TILENAME}.tif"
+    create_class_raster_raw(
+        in_points=IN_POINTS, output_file=str(output_file), res=1, raster_driver=RASTER_DRIVER, no_data_value=-9999.0
     )
-    assert raster == PATH1_EXPECTED  # good output filename
-    assert os.path.isfile(PATH1_EXPECTED)  # file exists
+    with rasterio.open(output_file) as raster:
+        band1 = raster.read(1)  # min
+        band2 = raster.read(2)  # max
+        band3 = raster.read(3)  # mean
+        band4 = raster.read(4)  # idw
+        band5 = raster.read(5)  # count
+        band6 = raster.read(6)  # stdev
+
+        assert band1[6, 17] == -9999.0
+        assert band2[6, 17] == -9999.0
+        assert band3[6, 17] == -9999.0
+        assert band4[6, 17] == -9999.0
+        assert band5[6, 17] == 0  # pixel with 0 point
+        assert band6[6, 17] == -9999.0
+
+        assert band1[0, 8] == -9999.0
+        assert band2[0, 8] == -9999.0
+        assert band3[0, 8] == -9999.0
+        assert band4[0, 8] == -9999.0
+        assert band5[0, 8] == 0  # pixel with 0 point
+        assert band6[0, 8] == -9999.0
+
+        assert band1[0, 10] == 2
+        assert band2[0, 10] == 2
+        assert band3[0, 10] == 2
+        assert band4[0, 10] == 2
+        assert band5[0, 10] == 1  # pixel with one single point (class 2)
+        assert band6[0, 10] == 0
+
+        assert band1[17, 14] == 1
+        assert band2[17, 14] == 65
+        assert round(band3[17, 14], 4) == 2.9375
+        assert round(band4[17, 14], 4) == 4.6058
+        assert band5[17, 14] == 64  # pixel with 64 points and at least class 1 and 65
+        assert round(band6[17, 14], 4) == 7.8220
 
 
-def test_step2_create_raster_fillgap():
-    raster_brut = step1_create_raster_brut(
-        in_points=IN_POINTS,
-        output_dir=OUTPUT_DIR,
-        output_filename=FILENAME,
-        res=1,
-        i=1,
-        output_extension=CONFIG.io.extension,
+def test_fill_gaps_raster():
+    input_file = Path("data") / "raster" / "class_raw" / f"{TILENAME}.tif"
+    output_file = Path(OUTPUT_DIR) / "fill_gaps_raster" / f"{TILENAME}.tif"
+    fill_gaps_raster(
+        in_raster=input_file,
+        output_file=output_file,
         raster_driver=RASTER_DRIVER,
     )
-    raster_fillgap = step2_create_raster_fillgap(
-        in_raster=raster_brut,
-        output_dir=OUTPUT_DIR,
-        output_filename=FILENAME,
-        output_extension=CONFIG.io.extension,
-        i=1,
-        raster_driver=RASTER_DRIVER,
-    )
-    assert raster_fillgap == PATH2_EXPECTED  # good output filename
-    assert os.path.isfile(PATH2_EXPECTED)  # file exists
+    with rasterio.open(output_file) as raster:
+        band1 = raster.read(1)
+        assert band1[6, 17] == -9999.0  # gap not filled
+        assert band1[0, 8] != -9999.0  # gap filled
+        assert band1[0, 10] == 2  # max (band2) kept from input raster
+        assert band1[17, 14] == 65  # max (band2) kept from input raster
 
 
-def test_step3_color_raster():
-    raster_brut = step1_create_raster_brut(
-        in_points=IN_POINTS,
-        output_dir=OUTPUT_DIR,
-        output_filename=FILENAME,
-        res=1,
-        i=1,
-        output_extension=CONFIG.io.extension,
-        raster_driver=RASTER_DRIVER,
+def test_add_color_to_raster():
+    input_file = Path("data") / "raster" / "class_fill_gaps" / f"{TILENAME}.tif"
+    output_file = Path(OUTPUT_DIR) / "add_color_to_raster" / f"{TILENAME}.tif"
+    add_color_to_raster(
+        in_raster=input_file,
+        output_file=output_file,
+        LUT=os.path.join("LUT", "LUT_CLASS.txt"),
     )
-    raster_color = step3_color_raster(
-        in_raster=raster_brut,
-        output_dir=OUTPUT_DIR,
-        tilename=FILENAME,
-        output_extension=CONFIG.io.extension,
-        verbose=VERBOSE,
-        i=1,
-        LUT=LUT_FILE,
-    )
-    assert raster_color == PATH3_EXPECTED
-    with rasterio.open(PATH3_EXPECTED) as raster:
+    with rasterio.open(output_file) as raster:
         band1 = raster.read(1)
         band2 = raster.read(2)
         band3 = raster.read(3)
-        assert band1[0, 0] == 255
-        assert band2[0, 0] == 128
-        assert band3[0, 0] == 0
+
+        assert band1[6, 17] == 255  # nodata -> color undefine in LUT => white
+        assert band2[6, 17] == 255  # nodata -> color undefine in LUT => white
+        assert band3[6, 17] == 255  # nodata -> color undefine in LUT => white
+
+        assert band1[0, 10] == 255  # sol (classe 2) -> color define in LUT
+        assert band2[0, 10] == 128  # sol (classe 2) -> color define in LUT
+        assert band3[0, 10] == 0  # sol (classe 2) -> color define in LUT
+
+        assert band1[17, 14] == 64  # artefact (classe 65) -> color define in LUT
+        assert band2[17, 14] == 0  # artefact (classe 65) -> color define in LUT
+        assert band3[17, 14] == 128  # artefact (classe 65) -> color define in LUT
+
+
+def test_create_map_class_raster_with_postprocessing_color_and_hillshade_default():
+    output_dir = Path(OUTPUT_DIR) / "create_map_class_raster_with_postprocessing_color_and_hillshade_default"
+    las_bounds = utils_pdal.get_bounds_from_las(INPUT_FILE)
+    with initialize(version_base="1.2", config_path="../configs"):
+        cfg = compose(
+            config_name="config_ctview",
+            overrides=[
+                f"io.output_dir={output_dir}",
+                f"io.tile_geometry.tile_coord_scale={TILE_COORD_SCALE}",
+                f"io.tile_geometry.tile_width={TILE_WIDTH}",
+                f"buffer.size={BUFFER_SIZE}",
+            ],
+        )
+    create_map_class_raster_with_postprocessing_color_and_hillshade(
+        INPUT_FILE, TILENAME, cfg.class_map, cfg.io, las_bounds
+    )
+    assert os.listdir(output_dir) == ["CLASS_FINAL"]
+    assert not glob.glob("tmp/tmp_class*")
+    with rasterio.Env():
+        with rasterio.open(Path(output_dir) / "CLASS_FINAL" / f"{TILENAME}_fusion_DSM_class.tif") as raster:
+            band1 = raster.read(1)
+            band2 = raster.read(2)
+            band3 = raster.read(3)
+            assert band1[0, 0] is not None
+            assert band2[0, 0] is not None
+            assert band3[0, 0] is not None
+            assert raster.res == (0.5, 0.5)
+
+
+def test_create_map_class_raster_with_postprocessing_color_and_hillshade_and_intermediate_files():
+    output_dir = (
+        Path(OUTPUT_DIR) / "create_map_class_raster_with_postprocessing_color_and_hillshade_and_intermediate_files"
+    )
+    las_bounds = utils_pdal.get_bounds_from_las(INPUT_FILE)
+    with initialize(version_base="1.2", config_path="../configs"):
+        # config is relative to a module
+        cfg = compose(
+            config_name="config_ctview",
+            overrides=[
+                f"io.output_dir={output_dir}",
+                f"io.tile_geometry.tile_coord_scale={TILE_COORD_SCALE}",
+                f"io.tile_geometry.tile_width={TILE_WIDTH}",
+                f"buffer.size={BUFFER_SIZE}",
+                "class_map.intermediate_dirs.CC_raw=CC_1_raw",
+                "class_map.intermediate_dirs.CC_raw_color=CC_2_bcolor",
+                "class_map.intermediate_dirs.CC_fillgap=CC_3_fg",
+                "class_map.intermediate_dirs.CC_fillgap_color=CC_4_fgcolor",
+                "class_map.intermediate_dirs.CC_crop=CC_5_crop",
+                "class_map.intermediate_dirs.dxm_raw=DSM_VAL",
+                "class_map.intermediate_dirs.dxm_hillshade=DSM_HS",
+            ],
+        )
+    create_map_class_raster_with_postprocessing_color_and_hillshade(
+        INPUT_FILE, TILENAME, cfg.class_map, cfg.io, las_bounds
+    )
+    assert set(os.listdir(output_dir)) == {
+        "CC_1_raw",
+        "CC_2_bcolor",
+        "CC_3_fg",
+        "CC_4_fgcolor",
+        "CC_5_crop",
+        "CLASS_FINAL",
+        "DSM_VAL",
+        "DSM_HS",
+    }
+
+    with rasterio.Env():
+        with rasterio.open(Path(output_dir) / "CC_4_fgcolor" / f"{TILENAME}_fillgap_color.tif") as raster:
+            band1 = raster.read(1)
+            band2 = raster.read(2)
+            band3 = raster.read(3)
+            assert band1[0, 0] == 255  # ground point
+            assert band2[0, 0] == 128
+            assert band3[0, 0] == 0
+            assert raster.res == (0.5, 0.5)
+        with rasterio.open(Path(output_dir) / "CLASS_FINAL" / f"{TILENAME}_fusion_DSM_class.tif") as raster:
+            band1 = raster.read(1)
+            band2 = raster.read(2)
+            band3 = raster.read(3)
+            assert band1[0, 0] is not None
+            assert band2[0, 0] is not None
+            assert band3[0, 0] is not None
+            assert raster.res == (0.5, 0.5)
 
 
 def test_main():
@@ -133,79 +211,19 @@ def test_main():
 
 
 def execute_main_default():
+    output_dir = Path(OUTPUT_DIR) / "execute_main_default"
     os.system(
         f"""
         python -m ctview.map_class \
         io.input_filename={INPUT_FILENAME} \
         io.input_dir={INPUT_DIR} \
-        io.output_dir={OUTPUT_DIR} \
+        io.output_dir={output_dir} \
+        io.tile_geometry.tile_coord_scale={TILE_COORD_SCALE} \
+        io.tile_geometry.tile_width={TILE_WIDTH} \
         """
     )
-    with rasterio.open(EXPECTED_FILLGAP_COLOR) as raster:
-        band1 = raster.read(1)
-        band2 = raster.read(2)
-        band3 = raster.read(3)
-        assert band1[0, 0] == 255
-        assert band2[0, 0] == 128
-        assert band3[0, 0] == 0
-        assert raster.res == (0.5, 0.5)
-
-
-def execute_main_change_pixel_size():
-    os.system(
-        f"""
-    python -m ctview.map_class \
-    io.input_filename={INPUT_FILENAME} \
-    io.input_dir={INPUT_DIR} \
-    io.output_dir={OUTPUT_DIR} \
-    class_map.pixel_size=5 \
-    """
-    )
-    with rasterio.open(EXPECTED_FILLGAP_COLOR) as raster:
-        band1 = raster.read(1)
-        band2 = raster.read(2)
-        band3 = raster.read(3)
-        assert band1[0, 0] == 255
-        assert band2[0, 0] == 128
-        assert band3[0, 0] == 0
-        assert raster.res == (5, 5)
-
-
-def test_create_map_class_raster_with_postprocessing_color_and_hillshade():
-    input_dir = Path("data") / "las" / "ground"
-    input_filename = "test_data_77055_627755_LA93_IGN69.las"
-    input_tilename = os.path.splitext(input_filename)[0]
-    tile_width = 50
-    tile_coord_scale = 10
-    buffer_size = 10
-    output_dir = Path(OUTPUT_DIR) / "create_map_class_raster_with_postprocessing_color_and_hillshade"
-    las_bounds = utils_pdal.get_bounds_from_las(os.path.join(input_dir, input_filename))
-    with initialize(version_base="1.2", config_path="../configs"):
-        # config is relative to a module
-        cfg = compose(
-            config_name="config_ctview",
-            overrides=[
-                f"io.input_filename={input_dir}",
-                f"io.input_dir={input_dir}",
-                f"io.output_dir={output_dir}",
-                f"io.tile_geometry.tile_coord_scale={tile_coord_scale}",
-                f"io.tile_geometry.tile_width={tile_width}",
-                f"buffer.size={buffer_size}",
-            ],
-        )
-    create_map_class_raster_with_postprocessing_color_and_hillshade(
-        os.path.join(input_dir, input_filename), input_tilename, cfg.class_map, cfg.io, las_bounds
-    )
     with rasterio.Env():
-        with rasterio.open(Path(output_dir) / "CC_4_fgcolor" / f"{input_tilename}_raster_fillgap_color.tif") as raster:
-            band1 = raster.read(1)
-            band2 = raster.read(2)
-            band3 = raster.read(3)
-            assert band1[0, 0] == 255
-            assert band2[0, 0] == 128
-            assert band3[0, 0] == 0
-            assert raster.res == (0.5, 0.5)
-        with rasterio.open(Path(output_dir) / "CC_6_fusion" / f"{input_tilename}_fusion_DSM_class.tif") as raster:
+        with rasterio.open(Path(output_dir) / "CLASS_FINAL" / f"{TILENAME}_fusion_DSM_class.tif") as raster:
             band1 = raster.read(1)
             band2 = raster.read(2)
             band3 = raster.read(3)
@@ -213,3 +231,27 @@ def test_create_map_class_raster_with_postprocessing_color_and_hillshade():
             assert band2[0, 0] is not None
             assert band3[0, 0] is not None
             assert raster.res == (0.5, 0.5)
+
+
+def execute_main_change_pixel_size():
+    output_dir = Path(OUTPUT_DIR) / "execute_main_change_pixel_size"
+    os.system(
+        f"""
+    python -m ctview.map_class \
+    io.input_filename={INPUT_FILENAME} \
+    io.input_dir={INPUT_DIR} \
+    io.output_dir={output_dir} \
+    io.tile_geometry.tile_coord_scale={TILE_COORD_SCALE} \
+    io.tile_geometry.tile_width={TILE_WIDTH} \
+    class_map.pixel_size=5 \
+    """
+    )
+    with rasterio.Env():
+        with rasterio.open(Path(output_dir) / "CLASS_FINAL" / f"{TILENAME}_fusion_DSM_class.tif") as raster:
+            band1 = raster.read(1)
+            band2 = raster.read(2)
+            band3 = raster.read(3)
+            assert band1[0, 0] is not None
+            assert band2[0, 0] is not None
+            assert band3[0, 0] is not None
+            assert raster.res == (5, 5)