+++ RELEASE_NOTES.md
@@ -1,7 +1,6 @@
 # dev
 - new:
-  - default config does not save intermediate files for dxm
-  - default config does not save intermediate files for density
+  - default config does not save intermediate files
 - dependencies: update pdal to 2.6+, python to 3.10+
 - refactor: 
   - simplify config file